refactor: break http into multiple modules
